industrial_automation/rpi5/requirements.txt
# Raspberry Pi 5 — runtime dependencies
# Install with: pip3 install -r requirements.txt

RPi.GPIO>=0.7.1
spidev>=3.6
